Broke up db_api.js into several model files
