+++ server/controllers/genreController.js
@@ -1,16 +1,17 @@
-var db = require(__baseDir + '/server/db_api/db_api.js');
+let GenreModel = require(__baseDir + '/server/db_api/GenreModel.js');
 
 module.exports = {
-    "index": function (req, res) {
+    index: function (req, res) {
         "use strict";
-        let send = [];
-        db.getAllGenres().then(genres => {
+        let responseArray = [];
+        GenreModel.getAll().then(genres => {
             if (genres) {
                 genres.forEach(genre => {
-                    send.push(genre[0]);
+                    responseArray.push(genre.name);
                 });
+
             }
-            res.send(send);
+            res.send(responseArray);
         });
     }
 };
